requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
requests==2.26.0
SQLAlchemy==1.4.28
urllib3==1.26.7
Werkzeug==2.0.2
python-dotenv==0.19.2
gunicorn==20.1.0
psycopg2-binary==2.9.2
